backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
groq==0.9.0
pydantic==2.7.1
pydantic-settings==2.3.1
python-dotenv==1.0.1
aiofiles==23.2.1
